web-proxy/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
pydantic